shipment product added

+++ src/main/java/com/fiveamazon/erp/service/impl/ShipmentServiceImpl.java
@@ -8,18 +8,24 @@ import com.fiveamazon.erp.common.SimpleConstant;
 import com.fiveamazon.erp.common.StatusConstant;
 import com.fiveamazon.erp.dto.*;
 import com.fiveamazon.erp.entity.*;
-import com.fiveamazon.erp.repository.ShipmentDetailRepository;
-import com.fiveamazon.erp.repository.ShipmentRepository;
-import com.fiveamazon.erp.repository.ShipmentViewRepository;
+import com.fiveamazon.erp.repository.*;
 import com.fiveamazon.erp.service.ShipmentService;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Sort;
+import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import javax.persistence.criteria.CriteriaBuilder;
+import javax.persistence.criteria.CriteriaQuery;
+import javax.persistence.criteria.Predicate;
+import javax.persistence.criteria.Root;
 import java.math.BigDecimal;
+import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
@@ -33,6 +39,8 @@ public class ShipmentServiceImpl implements ShipmentService {
     private ShipmentViewRepository shipmentViewRepository;
     @Autowired
     private ShipmentDetailRepository shipmentDetailRepository;
+    @Autowired
+    private ShipmentProductViewRepository shipmentProductViewRepository;
 
     @Override
     public Long countAll() {
@@ -67,6 +75,7 @@ public class ShipmentServiceImpl implements ShipmentService {
     @Override
     public ShipmentPO save(ShipmentDTO shipmentDTO) {
         Integer shipmentId = shipmentDTO.getId();
+        Integer excelId = shipmentDTO.getExcelId();
         ShipmentPO shipmentPO;
         if(shipmentId == null || shipmentId == 0){
             shipmentPO = new ShipmentPO();
@@ -76,13 +85,21 @@ public class ShipmentServiceImpl implements ShipmentService {
             shipmentPO = getById(shipmentId);
             shipmentPO.setUpdateDate(new Date());
             shipmentPO.setUpdateUser(shipmentDTO.getUsername());
-            shipmentDetailRepository.deleteByShipmentIdEqualsAndBoxNotLike(shipmentId, "Plan");
-            for(ShipmentDetailDTO shipmentDetailDTO: shipmentDTO.getShipmentDetailList()){
-                shipmentDetailDTO.setUsername(shipmentDTO.getUsername());
-                saveDetail(shipmentDetailDTO);
+            //if excel, the detail can't be updated, because it will delete sku
+            if(null == excelId || excelId == 0){
+                shipmentDetailRepository.deleteByShipmentIdEqualsAndBoxNotLike(shipmentId, "Plan");
+                for(ShipmentDetailDTO shipmentDetailDTO: shipmentDTO.getShipmentDetailList()){
+                    shipmentDetailDTO.setUsername(shipmentDTO.getUsername());
+                    saveDetail(shipmentDetailDTO);
+                }
             }
         }
-        BeanUtils.copyProperties(shipmentDTO, shipmentPO, "id");
+
+        if(null == excelId || excelId == 0){
+            BeanUtils.copyProperties(shipmentDTO, shipmentPO, "id");
+        }else{
+            BeanUtils.copyProperties(shipmentDTO, shipmentPO, "id", "weight");
+        }
         return save(shipmentPO);
     }
 
@@ -121,7 +138,11 @@ public class ShipmentServiceImpl implements ShipmentService {
         if(shipmentRepository.countByFbaNo(fbaNo) > 0){
             throw new SimpleCommonException("Duplicate FBA Found !");
         }
+        String excelDate = DateUtil.format(new Date(), "yyyyMMdd");
+        Integer excelId = uploadFbaDTO.getId();
         ShipmentPO shipmentPO = new ShipmentPO();
+        shipmentPO.setExcelDate(excelDate);
+        shipmentPO.setExcelId(excelId);
         shipmentPO.setBoxCount(uploadFbaDTO.getBoxCount());
         shipmentPO.setFbaNo(fbaNo);
         shipmentPO.setCreateDate(today);
@@ -207,6 +228,68 @@ public class ShipmentServiceImpl implements ShipmentService {
         shipmentDetailPO.setSkuId(boxJson.getInt("skuId"));
         shipmentDetailPO.setWeight(new BigDecimal(0));
         saveDetail(shipmentDetailPO);
+    }
+
+    @Override
+    public List<ShipmentProductViewPO> findAllProducts(ShipmentProductSearchDTO searchDTO) {
+        String dateFrom = searchDTO.getDateFrom().replaceAll("-", "");
+        String dateTo = searchDTO.getDateTo().replaceAll("-", "");
+        String dateType = searchDTO.getDateType();
+        String productId = searchDTO.getProductId();
+        String skuId = searchDTO.getSkuId();
+        Sort sort = Sort.by(Sort.Direction.DESC, "id");
+        PageRequest pageRequest = PageRequest.of(0, 100000000, sort);
+        Specification<ShipmentProductViewPO> specification = new Specification<ShipmentProductViewPO>() {
+            @Override
+            public Predicate toPredicate(Root root, CriteriaQuery criteriaQuery, CriteriaBuilder criteriaBuilder) {
+                List<Predicate> predicates = new ArrayList<>();
+                if (StringUtils.isNotEmpty(productId)) {
+                    predicates.add(criteriaBuilder.equal(root.get("productId"), productId));
+                }
+                if (StringUtils.isNotEmpty(skuId)) {
+                    predicates.add(criteriaBuilder.equal(root.get("skuId"), skuId));
+                }
+
+                switch (dateType){
+                    case "excelDate":
+                        if (StringUtils.isNotEmpty(dateFrom)) {
+                            predicates.add(criteriaBuilder.greaterThanOrEqualTo(root.get("excelDate"), dateFrom));
+                        }
+                        if (StringUtils.isNotEmpty(dateTo)) {
+                            predicates.add(criteriaBuilder.lessThanOrEqualTo(root.get("excelDate"), dateTo));
+                        }
+                        break;
+                    case "signedDate":
+                        if (StringUtils.isNotEmpty(dateFrom)) {
+                            predicates.add(criteriaBuilder.greaterThanOrEqualTo(root.get("signedDate"), dateFrom));
+                        }
+                        if (StringUtils.isNotEmpty(dateTo)) {
+                            predicates.add(criteriaBuilder.lessThanOrEqualTo(root.get("signedDate"), dateTo));
+                        }
+                        break;
+                    case "deliveryDate":
+                        if (StringUtils.isNotEmpty(dateFrom)) {
+                            predicates.add(criteriaBuilder.greaterThanOrEqualTo(root.get("deliveryDate"), dateFrom));
+                        }
+                        if (StringUtils.isNotEmpty(dateTo)) {
+                            predicates.add(criteriaBuilder.lessThanOrEqualTo(root.get("deliveryDate"), dateTo));
+                        }
+                        break;
+                    case "paymentDate":
+                        if (StringUtils.isNotEmpty(dateFrom)) {
+                            predicates.add(criteriaBuilder.greaterThanOrEqualTo(root.get("paymentDate"), dateFrom));
+                        }
+                        if (StringUtils.isNotEmpty(dateTo)) {
+                            predicates.add(criteriaBuilder.lessThanOrEqualTo(root.get("paymentDate"), dateTo));
+                        }
+                        break;
+                    default:
+                        break;
+                }
+                return criteriaBuilder.and(predicates.toArray(new Predicate[predicates.size()]));
+            }
+        };
+        return shipmentProductViewRepository.findAll(specification, pageRequest).getContent();
     }
